Move provenance reader class into separate file; remove unnecessary imports
