+++ src/lib/djerba/configure.py
@@ -1,17 +1,12 @@
 """Configure an INI file with Djerba inputs"""
 
-import configparser
-import csv
-import gzip
-import hashlib
 import logging
 import os
-import re
 from math import log2
 
+from djerba.provenance_reader import provenance_reader
 from djerba.sequenza import sequenza_reader, SequenzaError
 import djerba.util.constants as constants
-import djerba.util.provenance_index as index
 import djerba.util.ini_fields as ini
 from djerba.util.logger import logger
 from djerba.util.validator import path_validator
@@ -19,7 +14,7 @@ from djerba.util.validator import path_validator
 class configurer(logger):
     """
     Class to do configuration in main Djerba method
-    Discover and apply param updates to a ConfigParser object
+    Discover and apply param updates
     Param updates are automatically extracted from data sources, eg. file provenance
     """
 
@@ -212,281 +207,5 @@ class log_r_cutoff_finder:
         else:
             return log2(x)
 
-class provenance_reader(logger):
-
-    # internal dictionary keys
-    SAMPLE_NAME_KEY = 'sample_name'
-
-    # parent sample attribute keys
-    GEO_EXTERNAL_NAME = 'geo_external_name'
-    GEO_GROUP_ID = 'geo_group_id'
-    GEO_TISSUE_ORIGIN_ID = 'geo_tissue_origin'
-    GEO_TISSUE_TYPE_ID = 'geo_tissue_type'
-    GEO_TUBE_ID = 'geo_tube_id'
-
-    # TODO introduce a concept of 'sample name' (not just 'root sample name')
-    # allow user to specify sample names for WG/T, WG/N, WT
-    # use to disambiguate multiple samples from the same donor (eg. at different times)
-    # sanity checks on FPR results; if not OK, die with an informative error
-
-    def __init__(self, provenance_path, study, donor,
-                 wgn_sample=None, wgt_sample=None, wt_sample=None,
-                 log_level=logging.WARNING, log_path=None):
-        # get provenance for the study and donor
-        # if this proves to be too slow, can preprocess the file using zgrep
-        self.logger = self.get_logger(log_level, __name__, log_path)
-        self.logger.info("Reading provenance for study '%s' and donor '%s' " % (study, donor))
-        self.root_sample_name = donor
-        self.provenance = []
-        with gzip.open(provenance_path, 'rt') as infile:
-            reader = csv.reader(infile, delimiter="\t")
-            for row in reader:
-                if row[index.STUDY_TITLE] == study and \
-                   row[index.ROOT_SAMPLE_NAME] == self.root_sample_name and \
-                   row[index.SEQUENCER_RUN_PLATFORM_ID] != 'Illumina_MiSeq':
-                    self.provenance.append(row)
-        if len(self.provenance)==0:
-            # continue with empty provenance results, eg. for GSICAPBENCH testing
-            msg = "No provenance records found for study '%s' and donor '%s' " % (study, donor) +\
-                "in '%s'" % provenance_path
-            self.logger.warning(msg)
-            self.attributes = []
-            self.patient_id = None
-            self.tumour_id = None
-            self.normal_id = None
-        else:
-            self.logger.info("Found %d provenance records" % len(self.provenance))
-            # find relevant attributes from provenance
-            # start by finding sample_name/attributes and removing duplicates
-            provenance_subset = set()
-            for row in self.provenance:
-                columns = (
-                    row[index.SAMPLE_NAME],
-                    row[index.PARENT_SAMPLE_ATTRIBUTES]
-                )
-                provenance_subset.add(columns)
-            # parse the 'parent sample attributes' value and get a list of dictionaries
-            self.attributes = [self._parse_row_attributes(row) for row in provenance_subset]
-            self.patient_id = self._id_patient()
-            self.tumour_id = self._id_tumour()
-            self.normal_id = self._id_normal()
-
-    def _filter_rows(self, index, value, rows=None):
-        # find matching provenance rows from a list
-        if rows == None: rows = self.provenance
-        return filter(lambda x: x[index]==value, rows)
-
-    def _filter_metatype(self, metatype, rows=None):
-        return self._filter_rows(index.FILE_META_TYPE, metatype, rows)
-
-    def _filter_pattern(self, pattern, rows=None):
-        if rows == None: rows = self.provenance
-        return filter(lambda x: re.search(pattern, x[index.FILE_PATH]), rows)
-
-    def _filter_workflow(self, workflow, rows=None):
-        return self._filter_rows(index.WORKFLOW_NAME, workflow, rows)
-
-    def _get_most_recent_row(self, rows):
-        # if input is empty, raise an error
-        # otherwise, return the row with the most recent date field (last in lexical sort order)
-        # rows may be an iterator; if so, convert to a list
-        rows = list(rows)
-        if len(rows)==0:
-            msg = "Empty input to find most recent row; no rows meet filter criteria?"
-            self.logger.debug(msg)
-            raise MissingProvenanceError(msg)
-        else:
-            return sorted(rows, key=lambda row: row[index.LAST_MODIFIED], reverse=True)[0]
-
-    def _get_unique_value(self, key, check, reference=False):
-        """
-        Get unique value (if any) of key from self.attributes
-        Attributes is a list of dictionaries, each corresponding to one or more provenance rows
-        If check==True, check the tissue type ID to determine if the row refers to a reference (ie. normal)
-        If check==False, require a unique value across both tumour and normal (eg. to find the patient ID)
-        Then:
-        - If key is present, confirm that all members of the list have the same value
-        - If key does not exist for any member, return None
-        - If values for key are inconsistent, return None (additional error checking is done downstream)
-        """
-        value_set = set()
-        for row in self.attributes:
-            if check:
-                if reference and row.get(self.GEO_TISSUE_TYPE_ID)!='R':
-                    continue
-                elif not reference and row.get(self.GEO_TISSUE_TYPE_ID)=='R':
-                    continue
-            value = row.get(key)
-            if value:
-                value_set.add(value)
-            else:
-                self.logger.debug("No value found for {0} in {1}".format(key, row))
-        self.logger.debug("Candidate value set for key '{0}': {1}".format(key, value_set))
-        if len(value_set)==0:
-            self.logger.debug("No value found for {0}, reference = {1}".format(key, reference))
-            value = None
-        elif len(value_set)==1:
-            value = list(value_set).pop()
-        else:
-            msg = "Value for '{0}' with reference={1} is not unique: Found {2}".format(key, reference, value_set)
-            self.logger.debug(msg)
-            value = None
-        return value
-
-    def _id_normal(self):
-        self.logger.debug("Finding normal ID")
-        normal_id = self._id_tumour_normal(self.patient_id, reference=True)
-        self.logger.debug("Found normal ID: {0}".format(normal_id))
-        return normal_id
-
-    def _id_patient(self):
-        # parse the external name to get patient ID
-        patient_id_raw = self._get_unique_value(self.GEO_EXTERNAL_NAME, check=False)
-        if patient_id_raw == None:
-            msg = "Cannot initialize file provenance reader: No value found for metadata field '{0}'".format(self.GEO_EXTERNAL_NAME)
-            self.logger.error(msg)
-            raise RuntimeError(msg)
-        return re.split(',', patient_id_raw).pop(0)
-
-    def _id_tumour(self):
-        self.logger.debug("Finding tumour ID")
-        tumour_id = self._id_tumour_normal(self.patient_id, reference=False)
-        self.logger.debug("Found tumour ID: {0}".format(tumour_id))
-        return tumour_id
-
-    def _id_tumour_normal(self, patient_id, reference):
-        """
-        Find the tumour or normal ID -- process differs only by value of the 'reference' flag
-        tube ID > group_ID > constructed, in order of preference
-        """
-        tube_id = self._get_unique_value(self.GEO_TUBE_ID, check=True, reference=reference)
-        group_id = self._get_unique_value(self.GEO_GROUP_ID, check=True, reference=reference)
-        tissue_origin = self._get_unique_value(self.GEO_TISSUE_ORIGIN_ID, check=True, reference=reference)
-        if reference:
-            tissue_type = 'R'
-        else:
-            tissue_type = self._get_unique_value(self.GEO_TISSUE_TYPE_ID, check=True, reference=False)
-        if patient_id and tissue_origin and tissue_type:
-            constructed_id = "{0}_{1}_{2}".format(patient_id, tissue_origin, tissue_type)
-        else:
-            constructed_id = None
-        self.logger.debug("ID candidates: {0}, {1}, {2}".format(tube_id, group_id, constructed_id))
-        if tube_id:
-            chosen_id = tube_id
-            self.logger.debug("Using {0} value for ID: {1}".format(self.GEO_TUBE_ID, tube_id))
-        elif group_id:
-            msg = "Could not find {0}, using {1} for ID: {2}".format(self.GEO_TUBE_ID, self.GEO_GROUP_ID, group_id)
-            self.logger.warning(msg)
-            chosen_id = group_id
-        elif constructed_id:
-            msg = "Could not find {0} or {1}, constructing alternate ID: {2}".format(self.GEO_TUBE_ID, self.GEO_GROUP_ID, constructed_id)
-            self.logger.warning(msg)
-            chosen_id = constructed_id
-        else:
-            msg = "Unable to construct tumour/normal ID for patient ID '{0}'; specify manually in INI file".format(patient_id)
-            self.logger.error(msg)
-            raise RuntimeError(msg)
-        return chosen_id
-
-    def _parse_default(self, workflow, metatype, pattern):
-        # get most recent file of given workflow, metatype, and file path pattern
-        # self._filter_* functions return an iterator
-        iterrows = self._filter_workflow(workflow)
-        iterrows = self._filter_metatype(metatype, iterrows)
-        iterrows = self._filter_pattern(pattern, iterrows) # metatype usually suffices, but double-check
-        try:
-            row = self._get_most_recent_row(iterrows)
-            path = row[index.FILE_PATH]
-        except MissingProvenanceError as err:
-            msg = "No provenance records meet filter criteria: Workflow = {0}, ".format(workflow) +\
-                  "metatype = {0}, regex = {1}.".format(metatype, pattern)
-            self.logger.debug(msg)
-            path = None
-        return path
-
-    def _parse_row_attributes(self, row):
-        """
-        Input is a (sample name, attributes) pair from a provenance row
-        Parse the attributes string and return a dictionary
-        """
-        attrs = {}
-        attrs[self.SAMPLE_NAME_KEY] = row[0]
-        for entry in re.split(';', row[1]):
-            pair = re.split('=', entry)
-            if len(pair)!=2:
-                msg = "Expected attribute of the form KEY=VALUE, found '{0}'".format(entry)
-                self.logger.error(msg)
-                raise ValueError(msg)
-            attrs[pair[0]] = pair[1]
-        self.logger.debug("Found row attributes: {0}".format(attrs))
-        return attrs
-
-    def find_identifiers(self):
-        """
-        Find the tumour/normal/patient identifiers from file provenance
-        """
-        identifiers = {
-            ini.TUMOUR_ID: self.tumour_id,
-            ini.NORMAL_ID: self.normal_id,
-            ini.PATIENT_ID: self.patient_id
-        }
-        self.logger.debug("Found identifiers: {0}".format(identifiers))
-        return identifiers
-
-    def parse_arriba_path(self):
-        return self._parse_default('arriba', 'application/octet-stream', '\.fusions\.tsv$')
-
-    def parse_delly_path(self):
-        return self._parse_default('delly', 'application/vcf-gz', 'somatic_filtered\.delly\.merged\.vcf\.gz$')
-
-    def parse_gep_path(self):
-        return self._parse_default('rsem', 'application/octet-stream', '\.genes\.results$')
-
-    def parse_maf_path(self):
-        suffix = 'filter\.deduped\.realigned\.recalibrated\.mutect2\.filtered\.maf\.gz$'
-        return self._parse_default('variantEffectPredictor', 'application/txt-gz', suffix)
-
-    def parse_mavis_path(self):
-        return self._parse_default('mavis', 'application/zip-report-bundle', '(mavis-output|summary)\.zip$')
-
-    def parse_sequenza_path(self):
-        return self._parse_default('sequenza', 'application/zip-report-bundle', '_results\.zip$')
-
-    def parse_starfusion_predictions_path(self):
-        return self._parse_default('starFusion', 'application/octet-stream', 'star-fusion\.fusion_predictions\.tsv$')
-
-    def parse_wg_bam_path(self):
-        suffix = '{0}\.filter\.deduped\.realigned\.recalibrated\.bam$'.format(self.tumour_id)
-        return self._parse_default('bamMergePreprocessing', 'application/bam', suffix)
-
-    def parse_wg_bam_ref_path(self):
-        # find the reference (normal) BAM
-        suffix = '{0}\.filter\.deduped\.realigned\.recalibrated\.bam$'.format(self.normal_id)
-        return self._parse_default('bamMergePreprocessing', 'application/bam', suffix)
-
-    def parse_wg_index_path(self):
-        suffix = '{0}\.filter\.deduped\.realigned\.recalibrated\.bai$'.format(self.tumour_id)
-        return self._parse_default('bamMergePreprocessing', 'application/bam-index', suffix)
-
-    def parse_wg_index_ref_path(self):
-        # find the reference (normal) BAM index
-        suffix = '{0}\.filter\.deduped\.realigned\.recalibrated\.bai$'.format(self.normal_id)
-        return self._parse_default('bamMergePreprocessing', 'application/bam-index', suffix)
-
-    ### WT assay produces only 1 bam file; no need to consider tumour vs. reference
-
-    def parse_wt_bam_path(self):
-        # matches *Aligned.sortedByCoord.out.bam if *not* preceded by an index of the form ACGTACGT
-        suffix = '('+self.root_sample_name+'.+)((?<![ACGT]{8})\.Aligned)\.sortedByCoord\.out\.bam$'
-        return self._parse_default('STAR', 'application/bam', suffix)
-
-    def parse_wt_index_path(self):
-        # matches *Aligned.sortedByCoord.out.bam if *not* preceded by an index of the form ACGTACGT
-        suffix = '('+self.root_sample_name+'.+)((?<![ACGT]{8})\.Aligned)\.sortedByCoord\.out\.bai$'
-        return self._parse_default('STAR', 'application/bam-index', suffix)
-
 class MissingConfigError(Exception):
-    pass
-
-class MissingProvenanceError(Exception):
     pass
